edit.utils.remapparams: move the tests from the main section into edtf.tests
