+++ edtf/util.py
@@ -65,47 +65,3 @@ def remapparams(**remap):
         return remapparams_wrapper
 
     return remapparams_decorator
-
-
-if __name__ == "__main__":
-
-    @remapparams(parseAll="parse_all")
-    def parser(s, parse_all=True):
-        pass
-
-    assert parser.__name__ == "parser"  # noqa: S101
-    parser("foo")
-    # this should not warn
-    parser("foo", parse_all=False)
-    # this should warn, but only once
-    for _ in 1, 2:
-        parser("foo", parseAll=False)
-    try:
-        parser("foo", parseAll=False, parse_all=True)
-    except ValueError:
-        pass
-    else:
-        raise AssertionError("expected ValueError because of duplicated parameters")
-
-    try:
-
-        @remapparams()
-        def no_remappings():
-            pass
-    except ValueError:
-        pass
-    else:
-        raise AssertionError(
-            "expected ValueError from @remapparams() because no remappings"
-        )
-    try:
-
-        @remapparams(p1="p2", p2="p3")
-        def no_remappings():
-            pass
-    except ValueError:
-        pass
-    else:
-        raise AssertionError(
-            "expected ValueError from @remapparams() because p1 remaps to another remapped parameter"
-        )
